requirements.txt
tensorflow
numpy
matplotlib
flask
pillow
io
keras
sklearn